api/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
pydantic 
bcrypt 
python-jose